trading_bot/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alpaca-py==0.35.0
alpaca-trade-api==3.2.0
altgraph==0.17.4
annotated-types==0.7.0
arrow==1.3.0
attrs==24.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
briefcase==0.3.20
build==1.2.2.post1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
contourpy==1.3.1
cookiecutter==2.6.0
cycler==0.12.1
deprecation==2.1.0
distlib==0.3.9
dmgbuild==1.6.2
ds-store==1.3.1
filelock==3.16.1
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
mac-alias==2.2.2
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mplfinance==0.12.10b0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
numpy==2.2.0
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
psutil==6.1.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
soupsieve==2.6
sseclient-py==1.8.0
ta==0.11.0
text-unidecode==1.3
threadpoolctl==3.5.0
tomli_w==1.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
virtualenv==20.28.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wheel==0.45.1
yarl==1.18.3
yfinance==0.2.50
alpaca-py>=0.8.0
numpy>=1.21.0
pandas>=1.3.0
pyinstaller>=5.6.0
scikit-learn>=1.0.0
ta>=0.10.0
pytz>=2021.1
python-dotenv>=0.19.0